% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claudinLowData}
\alias{claudinLowData}
\title{claudinLowData for use in the claudinLow classifier. Data generously provided by Aleix Prat.}
\source{
http://jnci.oxfordjournals.org/cgi/content/full/98/4/262/DC1
}
\usage{
data(claudinLowData)
}
\description{
Training and Testing Data for use with the Claudin-Low Classifier
}
\details{
\itemize{
\item xd: Matrix of 807 features and 52 samples
\item classes: factor to split samples
\item nfeatures: number of features
\item nsamples: number of samples
\item fnames: names of features
\item snames: names of samples
}
}
\references{
Aleix Prat, Joel S Parker, Olga Karginova, Cheng Fan, Chad Livasy, Jason I Herschkowitz, Xiaping He, and Charles M. Perou (2010) "Phenotypic and molecular characterization of the claudin-low intrinsic subtype of breast cancer", Breast Cancer Research, 12(5):R68
}
\seealso{
\link{claudinLow}
}
\keyword{data}
