% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseIDs.R
\name{collapseIDs}
\alias{collapseIDs}
\title{Utility function to collapse IDs}
\usage{
collapseIDs(x,allids=row.names(x),method="mean")
}
\arguments{
\item{x}{Matrix of numbers.}

\item{allids}{Defaults to rownames of matrix.}

\item{method}{Default method is "mean".}
}
\value{
A matrix
}
\description{
Utility function called within the claudinLow classifier
}
\references{
citation("claudinLow")
}
\seealso{
\link{claudinLow}
}
