% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovcSigs.R
\name{.ovcSigs}
\alias{.ovcSigs}
\title{import utils}
\usage{
.ovcSigs(
  sigs = c("bentink2012_angiogenic", "crijns2009_sig", "yoshihara2010_sig",
    "spentzos2011_sig", "tcga2011_sig")
)
}
\description{
import utils
}
\keyword{internal}
