% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{genefu-package}
\alias{genefu}
\alias{genefu-package}
\title{genefu: Computation of Gene Expression-Based Signatures in Breast Cancer}
\description{
This package contains functions implementing various tasks usually required by gene expression analysis, especially in breast cancer studies: gene mapping between different microarray platforms, identification of molecular subtypes, implementation of published gene signatures, gene selection, and survival analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.pmgenomics.ca/bhklab/software/genefu}
}

}
\author{
\strong{Maintainer}: Benjamin Haibe-Kains \email{benjamin.haibe.kains@utoronto.ca}

Authors:
\itemize{
  \item Deena M.A. Gendoo
  \item Natchar Ratanasirigulchai
  \item Markus S. Schroeder
  \item Laia Pare
  \item Joel S Parker
  \item Aleix Prat
}

Other contributors:
\itemize{
  \item Nikta Feizi [contributor]
  \item Christopher Eeles [contributor]
}

}
\keyword{internal}
