\name{medianCtr}
\alias{medianCtr}
\title{Center around the median}
\description{
 Utility function called within the claudinLow classifier
}
\usage{
medianCtr(x)
}
\arguments{
  \item{x}{Matrix of numbers}
}
\value{
  Returns a matrix of median-centered numbers
}
\references{
  \code{citation("claudinLow")}
}
\seealso{ \code{\link[genefu]{claudinLow}} }


