% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapSets.R
\name{overlapSets}
\alias{overlapSets}
\title{Overlap two datasets}
\usage{
overlapSets(x,y)
}
\arguments{
\item{x}{Matrix1}

\item{y}{Matrix2}
}
\value{
A list of overlapped dataset
}
\description{
Utility function called within the claudinLow classifien.
}
\references{
citation("claudinLow")
}
\seealso{
\link{claudinLow}
}
