% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scmod1.robust}
\alias{scmod1.robust}
\title{Subtype Clustering Model using ESR1, ERBB2 and AURKA modules for identification of breast cancer molecular subtypes (Desmedt et al 2008)}
\format{
List of parameters for SCMOD1:
\itemize{
\item parameters: List of parameters for the mixture of three Gaussians (ER-/HER2-, HER2+ and ER+/HER2-) that define the Subtype Clustering Model. The structure is the same than for an \code{\link[mclust:Mclust]{mclust::Mclust()}} object.
\item cutoff.AURKA: Cutoff for AURKA module score in order to identify ER+/HER2- High Proliferation (aka Luminal B) tumors and ER+/HER2- Low Proliferation (aka Luminal A) tumors.
\item mod: ESR1, ERBB2 and AURKA modules.
}
}
\source{
\url{http://clincancerres.aacrjournals.org/content/14/16/5158.abstract?ck=nck}
}
\usage{
data(scmod1.robust)
}
\description{
List of parameters defining the Subtype Clustering Model as published in Desmedt et al 2008.
}
\references{
Desmedt C, Haibe-Kains B, Wirapati P, Buyse M, Larsimont D, Bontempi G, Delorenzi M, Piccart M, and Sotiriou C (2008) "Biological processes associated with breast cancer clinical outcome depend on the molecular subtypes", \emph{Clinical Cancer Research}, \emph{14}(16):5158--5165.
}
\keyword{data}
