% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig.gene70}
\alias{sig.gene70}
\title{Signature used to compute the 70 genes prognosis profile (GENE70) as published by van't Veer et al. 2002}
\format{
sig.gene70 is a matrix with 9 columns containing the annotations and information related to the signature itself.
}
\source{
\url{http://www.nature.com/nature/journal/v415/n6871/full/415530a.html}
}
\usage{
data(sig.gene70)
}
\description{
List of 70 agilent probe ids representing 56 unique genes included in the GENE70 signature. The EntrezGene.ID allows for mapping and the "average.good.prognosis.profile" values allows for signature computation.
}
\references{
L. J. van't Veer and H. Dai and M. J. van de Vijver and Y. D. He and A. A. Hart and M. Mao and H. L. Peterse and K. van der Kooy and M. J. Marton and A. T. Witteveen and G. J. Schreiber and R. M. Kerkhiven and C. Roberts and P. S. Linsley and R. Bernards and S. H. Friend (2002) "Gene Expression Profiling Predicts Clinical Outcome of Breast Cancer", Nature, 415:530--536.
}
\keyword{data}
