% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig.gene76}
\alias{sig.gene76}
\title{Signature used to compute the Relapse Score (GENE76) as published in Wang et al. 2005}
\format{
\code{sig.gene70} is a matrix with 10 columns containing the annotations and information related to the signature itself.
}
\source{
\url{http://www.thelancet.com/journals/lancet/article/PIIS0140-6736(05)17947-1/abstract}
}
\usage{
data(sig.gene76)
}
\description{
List of 76 affymetrix hgu133a probesets representing 60 unique genes included in the GENE76 signature. The EntrezGene.ID allows for mapping and the coefficient allows for signature computation.
}
\references{
Y. Wang and J. G. Klijn and Y. Zhang and A. M. Sieuwerts and M. P. Look and F. Yang and D. Talantov and M. Timmermans and M. E. Meijer-van Gelder and J. Yu and T. Jatkoe and E. M. Berns and D. Atkins and J. A. Foekens (2005) "Gene-Expression Profiles to Predict Distant Metastasis of Lymph-Node-Negative Primary Breast Cancer", Lancet, 365(9460):671--679.
}
\keyword{data}
