% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig.genius}
\alias{sig.genius}
\title{Gene Expression progNostic Index Using Subtypes (GENIUS) as published by Haibe-Kains et al. 2010.}
\format{
\code{sig.genius} is a list a three subtype signatures.
}
\description{
List of three gene signatures which compose the Gene Expression progNostic Index Using Subtypes (GENIUS) as published by Haibe-Kains et al. 2009. GENIUSM1, GENIUSM2 and GENIUSM3  are the ER-/HER2-, HER2+ and ER+/HER2- subtype signatures respectively.
}
\references{
Haibe-Kains B, Desmedt C, Rothe F, Sotiriou C and Bontempi G (2010) "A fuzzy gene expression-based computational approach improves breast cancer prognostication", Genome Biology, 11(2):R18
}
\keyword{data}
