% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig.tamr13}
\alias{sig.tamr13}
\title{Tamoxifen Resistance signature composed of 13 gene clusters (TAMR13) as published by Loi et al. 2008.}
\format{
sig.tamr13 is a list a 13 clusters of genes with their corresponding coefficient.
}
\usage{
data(sig.tamr13)
}
\description{
List of 13 clusters of genes (and annotations) and their corresponding coefficient as an additional attribute.
}
\references{
Loi S, Haibe-Kains B, Desmedt C, Wirapati P, Lallemand F, Tutt AM, Gillet C, Ellis P, Ryder K, Reid JF, Daidone MG, Pierotti MA, Berns EMJJ, Jansen MPHM, Foekens JA, Delorenzi M, Bontempi G, Piccart MJ and Sotiriou C (2008) "Predicting prognosis using molecular profiling in estrogen receptor-positive breast cancer treated with tamoxifen", BMC Genomics, 9(1):239
}
\keyword{data}
