% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigOvcAngiogenic}
\alias{sigOvcAngiogenic}
\title{sigOvcAngiogenic dataset}
\source{
\url{http://jnci.oxfordjournals.org/cgi/content/full/98/4/262/DC1}
}
\description{
sigOvcAngiogenic dataset
}
\references{
Bentink S, Haibe-Kains B, Risch T, Fan J-B, Hirsch MS, Holton K, Rubio R, April C, Chen J, Wickham-Garcia E, Liu J, Culhane AC, Drapkin R, Quackenbush JF, Matulonis UA (2012) "Angiogenic mRNA and microRNA Gene Expression Signature Predicts a Novel Subtype of Serous Ovarian Cancer", PloS one, 7(2):e30269
}
\keyword{data}
