% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigOvcCrijns}
\alias{sigOvcCrijns}
\title{sigOvcCrijns dataset}
\source{
\url{http://jnci.oxfordjournals.org/cgi/content/full/98/4/262/DC1}
}
\description{
sigOvcCrijns dataset
}
\references{
Crijns APG, Fehrmann RSN, de Jong S, Gerbens F, Meersma G J, Klip HG, Hollema H, Hofstra RMW, te Meerman GJ, de Vries EGE, van der Zee AGJ (2009) "Survival-Related Profile, Pathways, and Transcription Factors in Ovarian Cancer" PLoS Medicine, 6(2):e1000024.
}
\keyword{data}
