% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigOvcSpentzos}
\alias{sigOvcSpentzos}
\title{sigOcvSpentzos dataset}
\source{
\url{http://jnci.oxfordjournals.org/cgi/content/full/98/4/262/DC1}
}
\description{
sigOcvSpentzos dataset
}
\references{
Spentzos, D., Levine, D. A., Ramoni, M. F., Joseph, M., Gu, X., Boyd, J., et al. (2004). "Gene expression signature with independent prognostic significance in epithelial ovarian cancer". Journal of clinical oncology, 22(23), 4700--4710. doi:10.1200/JCO.2004.04.070
}
\keyword{data}
