% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigOvcTCGA}
\alias{sigOvcTCGA}
\title{sigOvcTCGA dataset}
\source{
\url{http://jnci.oxfordjournals.org/cgi/content/full/98/4/262/DC1}
}
\description{
sigOvcTCGA dataset
}
\references{
Bell D, Berchuck A, Birrer M et al. (2011) "Integrated genomic analyses of ovarian carcinoma", Nature, 474(7353):609--615
}
\keyword{data}
