% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigOvcYoshihara}
\alias{sigOvcYoshihara}
\title{sigOvcYoshihara dataset}
\source{
\url{http://jnci.oxfordjournals.org/cgi/content/full/98/4/262/DC1}
}
\description{
sigOvcYoshihara dataset
}
\references{
Yoshihara K, Tajima A, Yahata T, Kodama S, Fujiwara H, Suzuki M, Onishi Y, Hatae M, Sueyoshi K, Fujiwara H, Kudo, Yoshiki, Kotera K, Masuzaki H, Tashiro H, Katabuchi H, Inoue I, Tanaka K (2010) "Gene expression profile for predicting survival in advanced-stage serous ovarian cancer across two independent datasets", PloS one, 5(3):e9615.
}
\keyword{data}
