% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strescR.R
\name{strescR}
\alias{strescR}
\title{Utility function to escape LaTeX special characters
present in a string}
\usage{
strescR(strings)
}
\arguments{
\item{strings}{A vector of strings to deal with.}
}
\value{
A vector of strings with escaped characters within each string.
}
\description{
This function returns a vector of strings in which LaTeX special
characters are escaped, this was useful in conjunction with xtable.
}
\examples{
strescR("MISC_RNA")
strescR(c("BB_0001","BB_0002"))

}
\references{
citation("seqinr")
}
\seealso{
stresc
}
