% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamr13.R
\name{tamr13}
\alias{tamr13}
\title{Function to compute the risk scores of the tamoxifen resistance
signature (TAMR13)}
\usage{
tamr13(data, annot, do.mapping = FALSE, mapping, verbose = FALSE)
}
\arguments{
\item{data}{Matrix of gene expressions with samples in rows and probes
in columns, dimnames being properly defined.}

\item{annot}{Matrix of annotations with at least one column named
"EntrezGene.ID", dimnames being properly defined.}

\item{do.mapping}{TRUE if the mapping through Entrez Gene ids must be
performed (in case of ambiguities, the most variant probe is kept for
each gene), FALSE otherwise.}

\item{mapping}{Matrix with columns "EntrezGene.ID" and "probe" used to
force the mapping such that the probes are not selected based on their variance.}

\item{verbose}{TRUE to print informative messages, FALSE otherwise.}
}
\value{
A list with items:
\itemize{
\item score: Continuous signature scores.
\item risk: Binary risk classification, 1 being high risk and 0 being low
risk (not implemented, the function will return NA values).
}
}
\description{
This function computes signature scores from gene expression values
following the algorithm used for the Tamoxifen Resistance signature (TAMR13).
}
\examples{
# load TAMR13 signature
data(sig.tamr13)
# load VDX dataset
data(vdxs)
# compute relapse score
tamr13.vdxs <- tamr13(data=data.vdxs, annot=annot.vdxs, do.mapping=FALSE)
summary(tamr13.vdxs$score)

}
\references{
Loi S, Haibe-Kains B, Desmedt C, Wirapati P, Lallemand F, Tutt AM, Gillet C,
Ellis P, Ryder K, Reid JF, Daidone MG, Pierotti MA, Berns EMJJ, Jansen MPHM,
Foekens JA, Delorenzi M, Bontempi G, Piccart MJ and Sotiriou C (2008)
"Predicting prognosis using molecular profiling in estrogen receptor-
positive breast cancer treated with tamoxifen", BMC Genomics, 9(1):239
}
\seealso{
\link{gene76}
}
