\name{collapseIDs}
\alias{collapseIDs}
\title{Utility function to collapse IDs}
\description{
 Utility function called within the claudinLow classifier
}
\usage{
collapseIDs<-(x,allids=row.names(x),method="mean")
}

\references{
  \code{citation("claudinLow")}
}
\author{}
\seealso{ \code{\link[claudinLow]{claudinLow}} }


