\name{expos}
\alias{expos}
\alias{data.expos}
\alias{annot.expos}
\alias{demo.expos}
\docType{data}
\title{
Gene expression, annotations and clinical data from the International Genomics Consortium
}
\description{
This dataset contains (part of) the gene expression, annotations and clinical data from the expO dataset collected by the International Genomics Consortium (\url{http://www.intgen.org/expo/}).
}
\usage{data(expos)}
\format{
\code{\link[genefu]{expos}} is a dataset containing three matrices:
 \describe{
    \item{data.expos}{Matrix containing gene expressions as measured by Affymetrix hgu133plus2 technology (single-channel, oligonucleotides}
    \item{annot.expos}{Matrix containing annotations of ffymetrix hgu133plus2 microarray platform}
    \item{demo.expos}{Clinical information of the breast cancer patients whose tumors were hybridized}
}
}
\details{
This dataset has been generated by the International Genomics Consortium using Affymetrix hgu133plus2 technology. The gene expressions have been normalized using fRMA. Only part of the gene expressions (966) are contained in \code{\link[genefu]{data.expos}}.
}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2109}
}
\references{
International Genomics Consortium, \url{http://www.intgen.org}

McCall MN, Bolstad BM, Irizarry RA. (2010) "Frozen robust multiarray analysis (fRMA)", \emph{Biostatistics}, \bold{11}(2):242-253.
}
\examples{
data(expos)
}
\keyword{data}
