\name{genefu-package}
\alias{genefu-package}
\alias{genefu}
\docType{package}
\title{
Relevant Functions for Gene Expression Analysis, Especially in Breast Cancer.
}
\description{
This package contains functions implementing various tasks usually required by gene expression analysis, especially in breast cancer studies: gene mapping between different microarray platforms, identification of molecular subtypes, implementation of published gene signatures, gene selection, survival analysis, ...
}
\details{
\tabular{ll}{
Package: \tab genefu\cr
Type: \tab Package\cr
Version: \tab \Sexpr{packageDescription("genefu")$Version}\cr
Date: \tab \Sexpr{packageDescription("genefu")$Date}\cr
License: \tab Artistic-2.0\cr
}
}

\author{
\bold{Benjamin Haibe-Kains}

- Computational Biology and Functional Genomics, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://compbio.dfci.harvard.edu/}

- Center for Cancer Computational Biology, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://cccb.dfci.harvard.edu/index.html}

Former labs:

- Machine Learning Group (MLG), Universite Libre de Bruxelles, Bruxelles, Belgium 

\url{http://www.ulb.ac.be/di/mlg/}

- Breast Cancer Translational Laboratory (BCTL), Institut Jules Bordet, Bruxelles, Belgium 

\url{http://www.bordet.be/en/services/medical/array/practical.htm}


\bold{Maintainer}: \bold{Benjamin Haibe-Kains}

\email{bhaibeka@jimmy.harvard.edu}

\email{bhaibeka@ulb.ac.be}

\bold{Markus Schroeder}

\email{mschroed@jimmy.harvard.edu}
}
%\references{}
\keyword{ clustering }
\keyword{ models }
\keyword{ breast cancer }
\keyword{ prognosis }
\seealso{
\code{survcomp}
}
