\name{genefu-package}
\alias{genefu-package}
\alias{genefu}
\docType{package}
\title{
Relevant Functions for Gene Expression Analysis, Especially in Breast Cancer.
}
\description{
Various functions relevant for gene expression analysis with emphasis on breast cancer.
}
\details{
\tabular{ll}{
Package: \tab genefu\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2010-09-09\cr
License: \tab GPL-3\cr
}
}

\author{
Benjamin Haibe-Kains

- Computational Biology and Functional Genomics, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://compbio.dfci.harvard.edu/}

- Center for Cancer Computational Biology, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://cccb.dfci.harvard.edu/index.html}

Former labs:

- Machine Learning Group (MLG), Universite Libre de Bruxelles, Bruxelles, Belgium 

\url{http://www.ulb.ac.be/di/mlg/}

- Breast Cancer Translational Laboratory (BCTL), Institut Jules Bordet, Bruxelles, Belgium 

\url{http://www.bordet.be/en/services/medical/array/practical.htm}


\bold{Maintainer}: Benjamin Haibe-Kains

\email{bhaibeka@jimmy.harvard.edu}

\email{bhaibeka@ulb.ac.be}
}
%\references{}
\keyword{ clustering }
\keyword{ models }
\keyword{ breast cancer }
\keyword{ prognosis }
\seealso{
\code{survcomp}
}
