% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianCtr.R
\name{medianCtr}
\alias{medianCtr}
\title{Center around the median}
\usage{
medianCtr(x)
}
\arguments{
\item{x}{Matrix of numbers}
}
\value{
A matrix of median-centered numbers
}
\description{
Utility function called within the claudinLow classifier
}
\references{
citation("claudinLow")
}
\seealso{
\link{claudinLow}
}
