\name{overlapSets}
\alias{overlapSets}
\title{Overlap two datasets}
\description{
 Utility function called within the claudinLow classifier
}
\usage{
overlapSets(mat1,mat2)
}
\arguments{
  \item{mat1}{Matrix1}
  \item{mat2}{Matrix2}
}
\value{
  Overlapped dataset
}
\references{
  \code{citation("claudinLow")}
}
\author{}
\seealso{ \code{\link[claudinLow]{claudinLow}} }


