\name{overlapSets}
\alias{overlapSets}
\title{Overlap two datasets}
\description{
 Utility function called within the claudinLow classifier
}
\usage{
overlapSets(x,y)
}
\arguments{
  \item{x}{Matrix1}
  \item{y}{Matrix2}
}
\value{
  Overlapped dataset
}
\references{
  \code{citation("claudinLow")}
}
\seealso{ \code{\link[genefu]{claudinLow}} }


